+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.0] - 2021-01-11
+### Added
+- Number of files option
+### Changed
+- All flags type changed to `bool`
+- Menu optimized
+- The `logger` function enhanced.
+- Time format in the `logger` changed to `%Y-%m-%d %H:%M:%S`
+- `dl_maker` function modified
+- `tgf_maker` function modified
+- `gdf_maker` function modified
+- `run` function modified
 ## [0.9] - 2020-10-07
 ### Added
 - GEXF format
@@ -101,7 +113,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - JSON format
 - README
 
-[Unreleased]: https://github.com/sepandhaghighi/pyrgg/compare/v0.9...dev
+[Unreleased]: https://github.com/sepandhaghighi/pyrgg/compare/v1.0...dev
+[1.0]: https://github.com/sepandhaghighi/pyrgg/compare/v0.9...v1.0
 [0.9]: https://github.com/sepandhaghighi/pyrgg/compare/v0.8...v0.9
 [0.8]: https://github.com/sepandhaghighi/pyrgg/compare/v0.7...v0.8
 [0.7]: https://github.com/sepandhaghighi/pyrgg/compare/v0.6...v0.7